+++ vault/__init__.py
@@ -3,11 +3,13 @@ This is just a special file that tells pip that your main module is in this fold
 No need to add anything here. Feel free to delete this line when you make your own package
 Leave it empty
 """
+from .client import VaultClient
 from .kv2_engine import KV2Engine
 from .db_engine import DBEngine
 from .exceptions import WrongKV2Configuration, WrongDBConfiguration
 
 __all__ = [
+    'VaultClient',
     'KV2Engine',
     'DBEngine',
     'WrongKV2Configuration',