+++ CHANGELOG.md
@@ -3,11 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## v2.1.0 - 2024-05-30
+## v2.1.0 - 2024-06-03
 ### What's Changed
 **Full Changelog**: https://github.com/obervinov/vault-package/compare/v2.0.4...v2.1.0 by @obervinov in https://github.com/obervinov/vault-package/pull/47
 #### 💥 Breaking Changes
-* Methods `get_env()`, `prepare_client_configurator()`, `prepare_client_secrets()` and `init_instance()` of class `VaultClient()` have been privatized
+* Changed environment variables
+* Changed general class `VaultClient()` incoming parameters
+* Decomposition of one overloaded class into several smaller classes responsible for their own aspect of interaction with vault.
 #### 📚 Documentation
 * Update `*.md` templates for issues and pull requests
 #### 🚀 Features